requirements.txt
asgiref==3.7.2
Django==4.2
django-taggit==3.0.0
Markdown==3.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==4.2
django-taggit==3.0.0
Markdown==3.4.1
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==4.2
django-taggit==3.0.0
Markdown==3.4.1
psycopg==3.1.15
psycopg-binary==3.1.15
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==4.2.2
django-taggit==3.0.0
Markdown==3.4.1
psycopg==3.1.15
psycopg-binary==3.1.15
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.2
django-bootstrap5==23.3
django-filter==23.5
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.14.0
idna==3.6
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
psycopg==3.1.15
psycopg-binary==3.1.15
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
webencodings==0.5.1
asgiref==3.7.2
attrs==23.1.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.2
django-bootstrap5==23.3
django-filter==23.5
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
psycopg==3.1.15
psycopg-binary==3.1.15
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
asgiref==3.7.2
attrs==23.1.0
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.2
django-bootstrap5==23.3
django-filter==23.5
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.2.0
psycopg==3.1.15
psycopg-binary==3.1.15
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
